+++ nallely/osc_bus.py
@@ -1,13 +1,13 @@
-from collections import defaultdict
 import json
 import math
+from collections import defaultdict
+
 from pythonosc.dispatcher import Dispatcher
-from pythonosc.osc_server import ThreadingOSCUDPServer, BlockingOSCUDPServer
+from pythonosc.osc_server import BlockingOSCUDPServer, ThreadingOSCUDPServer
 from pythonosc.udp_client import SimpleUDPClient
 
-from .core.world import ThreadContext
-
 from .core.virtual_device import VirtualDevice, VirtualParameter
+from .core.world import ThreadContext
 from .websocket_bus import WebSocketBus